Rename src into pt_bot
