+++ pt_bot/calendars/__init__.py
@@ -1,6 +1,6 @@
 from aiogram import Router
 
-from bot.settings import Settings
+from pt_bot.settings import Settings
 
 from .containers import CalendarContainer
 from .routers import router
