+++ pt_bot/calendars/constants/calendar_category.py
