+++ pt_bot/calendars/constants/messages.py
