+++ pt_bot/main.py
@@ -4,11 +4,11 @@ from aiogram import Bot, Dispatcher
 from aiogram.client.default import DefaultBotProperties
 from aiogram.enums.parse_mode import ParseMode
 
-from bot.bot.commands import COMMANDS
-from bot.bot.containers import BotContainer
-from bot.calendars import get_router as get_calendar_router
-from bot.settings import Settings
-from bot.start import get_router as get_start_router
+from pt_bot.bot.commands import COMMANDS
+from pt_bot.bot.containers import BotContainer
+from pt_bot.calendars import get_router as get_calendar_router
+from pt_bot.settings import Settings
+from pt_bot.start import get_router as get_start_router
 
 
 async def _include_routers(dp: Dispatcher, settings: Settings) -> None:
