+++ pt_bot/start/containers.py
@@ -2,7 +2,7 @@ from os import path
 
 from dependency_injector import containers, providers
 
-from bot.core.db.utils import init_db_connection_pool, load_queries
+from pt_bot.core.db.utils import init_db_connection_pool, load_queries
 
 from .db.queries.builders import UserQueryBuilder
 from .db.repositories import UserRepository
@@ -12,7 +12,7 @@ from .services.users import UserService
 class StartContainer(containers.DeclarativeContainer):
     wiring_config = containers.WiringConfiguration(
         modules=[
-            "bot.start.routers",
+            "pt_bot.start.routers",
         ],
     )
 
@@ -20,7 +20,9 @@ class StartContainer(containers.DeclarativeContainer):
 
     user_unloaded_qb = providers.Singleton(UserQueryBuilder)
 
-    _user_queries_path = providers.Callable(path.join, env.project_dir, "src/start/db/queries/sql/user.sql")
+    _user_queries_path = providers.Callable(
+        path.join, env.project_dir, "pt_bot/start/db/queries/sql/user.sql"
+    )
     user_qb = providers.Callable(
         load_queries,
         builder=user_unloaded_qb,
