+++ run_app.sh
@@ -4,4 +4,4 @@ uv run alembic upgrade head
 echo -e "\n\nLoading reference data.\n\n"
 uv run python3 data/load.py  --file data/sql/reference_data.sql
 
-uv run python src/main.py
+uv run python pt_bot/main.py
