+++ secrets/.env.example
@@ -17,3 +17,8 @@ POSTGRES_DB=
 POSTGRES_HOST=
 POSTGRES_PORT=
 DB_DSN="postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:${POSTGRES_PORT}/${POSTGRES_DB}"
+
+# Prefect
+PREFECT_SERVER_API_HOST=0.0.0.0
+PREFECT_API_URL=http://prefect-server:4200/api
+PREFECT_API_DATABASE_CONNECTION_URL="postgresql+asyncpg://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${POSTGRES_HOST}:${POSTGRES_PORT}/${POSTGRES_DB}"
