+++ tests/unit/start/conftest.py
@@ -1,8 +1,8 @@
 import pytest
 
-from bot.settings import Settings
-from bot.start.containers import StartContainer
-from bot.start.services.users import UserService
+from pt_bot.settings import Settings
+from pt_bot.start.containers import StartContainer
+from pt_bot.start.services.users import UserService
 
 from .mocks.repositories import UserRepositoryMock
