+++ tests/unit/start/mocks/repositories.py
@@ -1,4 +1,4 @@
-from bot.start.errors import UserAlreadyExistsError
+from pt_bot.start.errors import UserAlreadyExistsError
 
 
 class UserRepositoryMock:
