+++ tests/unit/start/test_user_service.py
@@ -1,6 +1,6 @@
 from aiogram.types import User
 
-from bot.start.services.users import UserService
+from pt_bot.start.services.users import UserService
 
 
 class TestUserService:
